# Conflicts:
#	app/src/main/java/org/d3kad3nt/sunriseClock/ui/light/LightsFragment.java
#	app/src/main/java/org/d3kad3nt/sunriseClock/ui/light/lightDetail/LightDetailViewModel.java

+++ app/src/main/java/org/d3kad3nt/sunriseClock/data/model/endpoint/LightEndpoint.java
@@ -17,5 +17,5 @@ public interface LightEndpoint {
 
     LiveData<ApiResponse<ResponseBody>> setOnState(String id, boolean newState);
 
-    LiveData<ApiResponse<ResponseBody>> setBrightness(String endpointLightId, double brightness);
+    LiveData<ApiResponse<ResponseBody>> setBrightness(String endpointLightId, int brightness);
 }
